refactor(vscode): extract apply-to-local and worktree diff workflows out of AgentManagerApp
